samples: peripheral: radio_test: Fix IRQ priorities
Zephyr has added static assert verifying that IRQ priority set does not
exceed the lowest IRQ priority available in the system. The radio_test
sample did not comply with the requirements. Fix this by setting the IRQ
priorities in the sample to the lowest possible value available in the
system.

Signed-off-by: Robert Lubos <robert.lubos@nordicsemi.no>